+++ apm-sniffer/optional-plugins/optional-spring-plugins/optional-spring-cloud/gateway-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/cloud/gateway/v412x/define/AbstractGatewayV412EnhancePluginDefine.java
@@ -15,36 +15,21 @@
  *  limitations under the License.
  */
 
-package org.apache.skywalking.apm.plugin.spring.cloud.gateway.v4x.define;
+package org.apache.skywalking.apm.plugin.spring.cloud.gateway.v412x.define;
 
-import org.apache.skywalking.apm.agent.core.plugin.WitnessMethod;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInstanceMethodsEnhancePluginDefine;
-
-import java.util.Collections;
-import java.util.List;
-
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.returns;
 
 /**
  * This abstract class defines the <code>witnessClasses()</code> method,
  * and other plugin define classes need to inherit from this class
  */
-public abstract class AbstractGatewayV4EnhancePluginDefine extends ClassInstanceMethodsEnhancePluginDefine {
+public abstract class AbstractGatewayV412EnhancePluginDefine extends ClassInstanceMethodsEnhancePluginDefine {
 
     /**
      * @since 4.0.0
      */
     @Override
     protected String[] witnessClasses() {
-        return new String[]{"org.springframework.cloud.gateway.filter.factory.cache.LocalResponseCacheProperties"};
-    }
-
-    @Override
-    protected List<WitnessMethod> witnessMethods() {
-        return Collections.singletonList(
-                new WitnessMethod("org.springframework.cloud.gateway.config.LocalResponseCacheAutoConfiguration",
-                        named("responseCacheSizeWeigher").
-                                and(returns(named("org.springframework.cloud.gateway.filter.factory.cache.ResponseCacheSizeWeigher")))));
+        return new String[]{"org.springframework.cloud.gateway.event.RouteDeletedEvent"};
     }
 }